+++ pyincore/models/hazard/earthquake.py
@@ -4,19 +4,19 @@
 # terms of the Mozilla Public License v2.0 which accompanies this distribution,
 # and is available at https://www.mozilla.org/en-US/MPL/2.0/
 from pyincore import HazardService, Dataset
-from pyincore.models.hazard import Hazard
-from pyincore.models.hazardDataset import HurricaneDataset
+from pyincore.models.hazard.hazard import Hazard
+from pyincore.models.hazard.hazardDataset import EarthquakeDataset
 
 
-class Hurricane(Hazard):
+class Earthquake(Hazard):
 
     def __init__(self, metadata):
         super().__init__(metadata)
         self.hazardDatasets = []
         if "hazardDatasets" in metadata:
             for hazardDataset in metadata["hazardDatasets"]:
-                self.hazardDatasets.append(HurricaneDataset(hazardDataset))
-        self.hazard_type = "hurricane"
+                self.hazardDatasets.append(EarthquakeDataset(hazardDataset))
+        self.hazard_type = "earthquake"
 
     @classmethod
     def from_hazard_service(cls, id: str, hazard_service: HazardService):
@@ -30,23 +30,22 @@ class Hurricane(Hazard):
             obj: Hazard from Data service.
 
         """
-        metadata = hazard_service.get_hurricane_metadata(id)
+        metadata = hazard_service.get_earthquake_hazard_metadata(id)
         instance = cls(metadata)
         return instance
 
     def read_hazard_values(self, payload: list, hazard_service=None, **kwargs):
-        """ Retrieve bulk hurricane hazard values either from the Hazard service or read it from local Dataset
+        """ Retrieve bulk earthquake hazard values either from the Hazard service or read it from local Dataset
 
         Args:
             payload (list):
             hazard_service (obj): Hazard service.
-            timeout (tuple): Timeout for the request.
             kwargs (dict): Keyword arguments.
         Returns:
             obj: Hazard values.
 
         """
         if self.id and self.id != "" and hazard_service is not None:
-            return hazard_service.post_hurricane_hazard_values(self.id, payload, **kwargs)
+            return hazard_service.post_earthquake_hazard_values(self.id, payload, **kwargs)
         else:
             return self.read_local_raster_hazard_values(payload)
