+++ pyincore/models/hazard/hazard.py
