+++ pyincore/models/hazard/hurricane.py
@@ -3,20 +3,20 @@
 # This program and the accompanying materials are made available under the
 # terms of the Mozilla Public License v2.0 which accompanies this distribution,
 # and is available at https://www.mozilla.org/en-US/MPL/2.0/
-from pyincore import HazardService, Dataset
-from pyincore.models.hazard import Hazard
-from pyincore.models.hazardDataset import FloodDataset
+from pyincore import HazardService
+from pyincore.models.hazard.hazard import Hazard
+from pyincore.models.hazard.hazardDataset import HurricaneDataset
 
 
-class Flood(Hazard):
+class Hurricane(Hazard):
 
     def __init__(self, metadata):
         super().__init__(metadata)
         self.hazardDatasets = []
         if "hazardDatasets" in metadata:
             for hazardDataset in metadata["hazardDatasets"]:
-                self.hazardDatasets.append(FloodDataset(hazardDataset))
-        self.hazard_type = "flood"
+                self.hazardDatasets.append(HurricaneDataset(hazardDataset))
+        self.hazard_type = "hurricane"
 
     @classmethod
     def from_hazard_service(cls, id: str, hazard_service: HazardService):
@@ -30,12 +30,12 @@ class Flood(Hazard):
             obj: Hazard from Data service.
 
         """
-        metadata = hazard_service.get_flood_metadata(id)
+        metadata = hazard_service.get_hurricane_metadata(id)
         instance = cls(metadata)
         return instance
 
     def read_hazard_values(self, payload: list, hazard_service=None, **kwargs):
-        """ Retrieve bulk flood hazard values either from the Hazard service or read it from local Dataset
+        """ Retrieve bulk hurricane hazard values either from the Hazard service or read it from local Dataset
 
         Args:
             payload (list):
@@ -47,6 +47,6 @@ class Flood(Hazard):
 
         """
         if self.id and self.id != "" and hazard_service is not None:
-            return hazard_service.post_flood_hazard_values(self.id, payload, **kwargs)
+            return hazard_service.post_hurricane_hazard_values(self.id, payload, **kwargs)
         else:
             return self.read_local_raster_hazard_values(payload)
